Calculate the optimal threshold
Calculate the optimal threshold for model.
To catch the frauds at a higher percentage to lessen economic loss for a company however we will have more wrong predictions on 0s, this will be result in extra work load for a company. It’s a good trade off for the company for them to work internally in the wrong predictions than to lose more financial.